% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_codons.R
\name{mk_codons}
\alias{mk_codons}
\title{DNA sequence to Codons}
\usage{
mk_codons(dna, s = 1)
}
\arguments{
\item{dna}{List of nucleotides (A,T,G,C)}
}
\value{
codons Triplets of nucleotides
}
\description{
Separates one sequence of aa into condons
}
\examples{
mk_codons("ATCGCTATG")

}
