% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_dna.R
\name{random_dna}
\alias{random_dna}
\title{Random DNA}
\usage{
random_dna(l)
}
\arguments{
\item{l}{lenght of DNA sequence}
}
\value{
dna The random DNA sequence
}
\description{
Creates a random sequence of DNA nucleotides
}
\examples{
random_dna(5)

}
