% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_rna.R
\name{random_rna}
\alias{random_rna}
\title{Get a random RNA sequence}
\usage{
random_rna(l)
}
\arguments{
\item{l}{Length of RNA sequence}
}
\value{
rna String of random RNA sequence
}
\description{
Get a random RNA sequence
}
